requirements.txt
python-dotenv
datasets 
trl 
peft 
bitsandbytes 
sentencepiece 
wandb
semantic-kernel
google-generativeai
beautifulsoup4
selenium
webdriver-manager
llama-index